requirements.txt
libevdev==0.12
MouseInfo==0.1.3
mss==10.0.0
numpy==2.3.1
pillow==11.2.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
vgamepad==0.1.0
